ISSUE 87 Added new field to critical writing field_cw_in_platform_software.
This is a new field to note platform that a piece of critical writing was created in. This is distinct from the field for a platform referenced by the critical writing. Note: I added this to features module info pages. Stein should check I didn't break anything and did this correctly.