issue 179 added filters and got them working on data export view.
Had to recreate the filters from scratch and make "Apply to All Views (except overidden)
and then override the block view. This appears to be the solution to the issue. Hopefully.
So the key is to recreate the filters on the page view as the exposed filters need to
originate there and be part of the display on the data export.