Issue 269 - Remove /node and change news path to /news
Drupal comes with a default front page example.com/node. ELMCIP
uses another front page. ELMCIP also uses a view that pulls out
the nodes that is "promoted" to front page (views_recent_posts)
that create a short list on the front page + exposes all story
types, promoted or not.

This remove /node and changes news to /news to prepare us to remove
the block from front page and replace it with a menu link.